pyproject.toml
[project]
name = "resume-critiquer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai",
    "pyarrow>=21.0.0",
    "pypdf",
    "python-dotenv>=1.1.1",
    "streamlit",
]
